backend/requirements.txt
-i https://pypi.org/simple
alembic==1.13.1 ; python_version >= '3.8'
blinker==1.7.0 ; python_version >= '3.8'
click==8.1.7 ; python_version >= '3.7'
flask==3.0.2
flask-cors==4.0.0
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.3 ; python_version >= '3.7'
mako==1.3.2 ; python_version >= '3.8'
markupsafe==2.1.5 ; python_version >= '3.7'
packaging==24.0 ; python_version >= '3.7'
python-dotenv==1.0.1
sqlalchemy==2.0.29 ; python_version >= '3.7'
sqlalchemy-serializer==1.4.1
typing-extensions==4.10.0 ; python_version >= '3.8'
werkzeug==3.0.1 ; python_version >= '3.8'
